added question tree data structure, and starting unit testing
