+++ app1.py
@@ -1,0 +1,119 @@
+# Python libraries that we need to import for our bot
+import random
+from flask import Flask, request
+from pymessenger.bot import Bot
+import os
+from QuestionsCollection import *
+from wit import Wit
+
+app = Flask(__name__)
+ACCESS_TOKEN = os.environ['ACCESS_TOKEN']
+VERIFY_TOKEN = os.environ['VERIFY_TOKEN']
+# WIT_ACCESS_TOKEN = os.environ['WIT_ACCESS_TOKEN']
+
+bot = Bot(ACCESS_TOKEN)
+#BLOCK
+# wit_client = Wit(WIT_ACCESS_TOKEN)
+# resp = wit_client.message("Hi there!")
+# print('Yay! got Wit.ai response: ' + str(resp))
+
+
+@app.route('/', methods=['GET', 'POST'])
+def receive_message():
+    if request.method == 'GET':
+        """Before allowing people to message your bot, Facebook has implemented a verify token
+        that confirms all requests that your bot receives came from Facebook."""
+        token_sent = request.args.get("hub.verify_token")
+        return verify_fb_token(token_sent)
+    # if the request was not get, it must be POST and we can just proceed with sending a message back to user
+    else:
+        # get whatever message a user sent the bot
+        output = request.get_json()
+        for event in output['entry']:
+            messaging = event['messaging']
+            for message in messaging:
+                if message.get('message'):
+                    #TODO: 1: Parsing the user message
+                    """
+                    1//
+                    THIS IS OUR PLAYBLOCK <- BLOCK CAN BE HERE
+                    what did he say, what it means and so on
+                    """
+                    # Facebook Messenger ID for user so we know where to send response back to
+                    recipient_id = message['sender']['id']
+                    #TODO: 2: Saving the user message in a dictionary somewhere
+                    """
+                    2//
+                    IN THIS BLOCK WE'D LIKE TO SEND THE MESSAGE DATA TO A DICTIONARY HOLDING ALL MESSAGES FROM THE SAME ID
+                    {SENDER_ID: CURRENT_MESSAGE}
+                    """
+
+                    print(message['message']['text'])
+                    quick_reply_ops = [{
+                        "content_type": "text",
+                        "title": "I am awesome",
+                        "payload": "I am more awesome",
+                    }, {
+                        "content_type": "text",
+                        "title": "I am awesome",
+                        "payload": "I am more awesome",
+                    }]
+                    print(message)
+                    print(message['message'])
+                    if message['message'].get('text'):
+                        response_sent_text = get_message()
+                        send_message(recipient_id, response_sent_text)
+                        # send_quick_reply(message, quick_reply_ops, recipient_id)
+                    # if user sends us a GIF, photo,video, or any other non-text item
+                    if message['message'].get('attachments'):
+                        response_sent_nontext = get_message()
+                        send_message(recipient_id, response_sent_nontext)
+                    #TODO: 3: Check if the user finished answering all of the questions and sign up
+                    """
+                    3//
+                    LAST BLOCK - USER HAS FINISHED ALL THE QUESTIONS -> SIGN HIM UP!
+                    """
+    return "Message Processed"
+
+
+def send_quick_reply(message, quick_reply, recipient_id):
+
+    payload = {
+        'recipient': {
+            'id': recipient_id
+        },
+        "messaging_type": "RESPONSE",
+        "message": {
+            'text': message
+        },
+        "quick_replies": quick_reply
+    }
+    return Bot.send_raw(payload)
+
+
+def verify_fb_token(token_sent):
+    # take token sent by facebook and verify it matches the verify token you sent
+    # if they match, allow the request, else return an error
+    if token_sent == VERIFY_TOKEN:
+        return request.args.get("hub.challenge")
+    return 'Invalid verification token'
+
+
+# chooses a random message to send to the user
+def get_message():
+    sample_responses = ["You are stunning!", "We're proud of you.", "Keep on being you!",
+                        "We're greatful to know you :)"]
+    # return selected item to the user
+    return random.choice(sample_responses)
+
+
+# uses PyMessenger to send response to user
+def send_message(recipient_id, response):
+    # sends user the text message provided via input response parameter
+    bot.send_text_message(recipient_id, response)
+    return "success"
+
+
+if __name__ == "__main__":
+    dict_of_users = {}
+    app.run()
